Moved Size and Die to monsters folder
